Arrangement for Nuget package - rearranged code directory, namespace and lib details for a test publication fo Nuget package.
