+++ sacco/Code/Models/Transactions/StdCoin.cs
@@ -9,12 +9,10 @@
 //
 using System;
 using System.Collections.Generic;
-using System.Text;
 using System.Diagnostics;
-using Newtonsoft.Json;
 
 
-namespace sacco.Lib
+namespace commercio.sacco.lib
 {
     public class StdCoin
     {
