+++ sacco/Code/Models/Transactions/StdFee.cs
@@ -9,11 +9,10 @@
 //
 using System;
 using System.Collections.Generic;
-using System.Text;
 using System.Diagnostics;
 
 
-namespace sacco.Lib
+namespace commercio.sacco.lib
 {
     public class StdFee
     {
