+++ sacco/Code/Models/Transactions/StdMsg.cs
@@ -11,12 +11,10 @@
 
 using System;
 using System.Collections.Generic;
-using System.Text;
 using System.Diagnostics;
-using Newtonsoft.Json;
 
 
-namespace sacco.Lib
+namespace commercio.sacco.lib
 {
     public class StdMsg
     {
