+++ sacco/Code/Models/Transactions/StdSignatureMessage.cs
@@ -8,12 +8,10 @@
 //
 using System;
 using System.Collections.Generic;
-using System.Text;
 using System.Diagnostics;
-using Newtonsoft.Json;
 
 
-namespace sacco.Lib
+namespace commercio.sacco.lib
 {
     public class StdSignatureMessage
     {
