+++ sacco/Code/Models/Transactions/StdTx.cs
@@ -8,12 +8,11 @@
 //
 using System;
 using System.Collections.Generic;
-using System.Text;
 using System.Diagnostics;
 using Newtonsoft.Json;
 
 
-namespace sacco.Lib
+namespace commercio.sacco.lib
 {
     public class StdTx
     {
