+++ sacco/Code/NetworkInfo.cs
@@ -12,7 +12,7 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 
-namespace sacco.Lib
+namespace commercio.sacco.lib
 {
     // *** This is inherited by Equatable in Dart Package!
     //  There is no such Class in C# - we include Compare-Net-Objects Nuget package for the purpose - see https://github.com/GregFinzer/Compare-Net-Objects
