+++ sacco/Code/Wallet.cs
@@ -15,9 +15,7 @@
 
 using System;
 using System.Collections.Generic;
-using System.Text;
 using System.Diagnostics;
-using System.IO;
 using NBitcoin;
 using Org.BouncyCastle.Asn1;
 using Org.BouncyCastle.Asn1.X9;
@@ -32,7 +30,7 @@ using Org.BouncyCastle.Crypto.Signers;
 
 
 
-namespace sacco.Lib
+namespace commercio.sacco.lib
 {
     // *** This is inherited by Equatable in Dart Package!
     //  There is no such Class in C# - we include Compare-Net-Objects Nuget package for the purpose - see https://github.com/GregFinzer/Compare-Net-Objects
