+++ sacco/Code/utils/AccountDataRetrieval.cs
@@ -15,9 +15,9 @@ using System.Threading.Tasks;
 using System.Net.Http;
 using Newtonsoft.Json.Linq;
 
-namespace sacco.Lib
+namespace commercio.sacco.lib
 {
-    public class NodeInfoRetrieval
+    public class AccountDataRetrieval
     {
         #region Instance Variables
 
@@ -37,11 +37,11 @@ namespace sacco.Lib
         /// Reads the account endpoint and retrieves data from it.
         /// *** This is ported from Dart Future to C# Task<TResult> 
         /// *** The usage of the class should be mantained - to be checked
-        public static async Task<NodeInfo> getNodeInfo(Wallet wallet)
+        public static async Task<AccountData> getAccountData(Wallet wallet)  
         {
             // Build the models.wallet api url
-            String endpoint = $"{wallet.networkInfo.lcdUrl}/node_info";
-
+            String endpoint = $"{wallet.networkInfo.lcdUrl}/auth/accounts/{wallet.bech32Address}";
+            
             // Get the server response
             HttpResponseMessage response = await client.GetAsync(endpoint);
             if (response.StatusCode != System.Net.HttpStatusCode.OK)
@@ -51,9 +51,8 @@ namespace sacco.Lib
             }
 
             // Parse the data
-            // Parse the data
             String jsonResponse = await response.Content.ReadAsStringAsync();
-            return (parseServerNodeInfo(jsonResponse));
+            return parseServerReturnedAccount(jsonResponse);
         }
 
 
@@ -62,30 +61,45 @@ namespace sacco.Lib
         #region Helpers
 
         // Moved the parsing code in a helper public method, so I can test it without net...
-        public static NodeInfo parseServerNodeInfo(String stringResponse)
+        public static AccountData parseServerReturnedAccount(String stringResponse)
         {
-            NodeInfo wkNode;
-            Dictionary<String, Object> jsonResponse, json = null;
-            String networkName = "";
+            AccountData wkAccount;
+            Dictionary<String, Object> jsonResponse, json = null, value = null;
+            List<StdCoin> coins = null;
+            String accountNumber = "", sequence = "";
             Object outValue;
 
             // Parse the data
             JObject jsonObj = JObject.Parse(stringResponse);
             jsonResponse = jsonObj.ToObject<Dictionary<String, Object>>();
-            // Get the "node_info" data from the response
-            if (jsonResponse.TryGetValue("node_info", out outValue))
+            // Get the "result" data from the response
+            if (jsonResponse.TryGetValue("result", out outValue))
             {
                 json = (outValue as JObject).ToObject<Dictionary<String, Object>>();
             }
-            // Get the "network" data from the node_info
-            if (json.TryGetValue("network", out outValue))
+            // Get the "value" data from the result
+            if (json.TryGetValue("value", out outValue))
             {
-                networkName = outValue as String;
+                value = (outValue as JObject).ToObject<Dictionary<String, Object>>();
             }
-            // Create the NodeInfo
-            wkNode = new NodeInfo(network: networkName);
+            // get various data from the value
+            if (value.TryGetValue("account_number", out outValue))
+            {
+                accountNumber = outValue as String;
+            }
+            if (value.TryGetValue("sequence", out outValue))
+            {
+                sequence = outValue as String;
+            }
+            // Get the coins - Careful, this is a List! so it's decoded as a JArray
+            if (value.TryGetValue("coins", out outValue))
+            {
+                coins = (outValue as JArray).ToObject<List<StdCoin>>();
+            }
+            // Create the account
+            wkAccount = new AccountData(accountNumber: accountNumber, sequence: sequence, coins: coins);
 
-            return (wkNode);
+            return (wkAccount);
         }
 
         #endregion
