+++ sacco/sacco.csproj
@@ -4,10 +4,17 @@
     <TargetFramework>netcoreapp2.1</TargetFramework>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <Company>Blockit S.r.l.</Company>
-    <Copyright>Copyright  2019 Commercio.network, Blockit S.r.l.</Copyright>
-    <Product>Commercio</Product>
-    <Description>Porting di Sacco in ambiente NET.Core</Description>
-    <RootNamespace>sacco.Lib</RootNamespace>
+    <Copyright>Copyright (c) 2019 Commercio.network, Blockit S.r.l.</Copyright>
+    <Product>Commercio.network</Product>
+    <Description>sacco.lib - core library for Commercio.network - NET.Core porting</Description>
+    <RootNamespace>commercio.sacco.lib</RootNamespace>
+    <Authors>Commercio.network, Blockit S.r.l.</Authors>
+    <PackageProjectUrl>https://github.com/commercionetwork</PackageProjectUrl>
+    <RepositoryUrl>https://github.com/commercionetwork/sacco.cs</RepositoryUrl>
+    <RepositoryType>GitHub</RepositoryType>
+    <Version>0.1.0</Version>
+    <AssemblyName>commercio.sacco.lib</AssemblyName>
+    <PackageReleaseNotes>First version in C#for .NETcore - porting from Dart version</PackageReleaseNotes>
   </PropertyGroup>
 
   <ItemGroup>
