+++ sacco_test/MainLib/Networkinfo_test.cs
@@ -2,7 +2,7 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Collections.Generic;
 using KellermanSoftware.CompareNetObjects;
-using sacco.Lib;
+using commercio.sacco.lib;
 
 
 namespace sacco_test
