+++ sacco_test/Models/AccountData_test.cs
@@ -2,7 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Text;
-using sacco.Lib;
+using commercio.sacco.lib;
 
 
 
@@ -32,7 +32,7 @@ namespace sacco_test
             // Copying here the string generated by Dart
             // Assert.AreEqual(s, "number: MyAccountNumber, sequence: MyAccountSequence, coins: [Instance of 'StdCoin', Instance of 'StdCoin']");
             // *** The string generated in the Dart environment is different fro the one in C# - Is this a problem?
-            Assert.AreEqual(s, "number: MyAccountNumber, sequence: MyAccountSequence, coins: System.Collections.Generic.List`1[sacco.Lib.StdCoin]");
+            Assert.AreEqual(s, "number: MyAccountNumber, sequence: MyAccountSequence, coins: System.Collections.Generic.List`1[commercio.sacco.lib.StdCoin]");
         }
     }
 }
