+++ sacco_test/utils/AccountDataRetriever_test.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 using Newtonsoft.Json;
-using sacco.Lib;
+using commercio.sacco.lib;
 
 namespace sacco_test
 {
@@ -42,7 +42,7 @@ namespace sacco_test
             // Copying here the string generated by Dart
             // Assert.AreEqual(s, "number: MyAccountNumber, sequence: MyAccountSequence, coins: [Instance of 'StdCoin', Instance of 'StdCoin']");
             // *** The string generated in the Dart environment is different fro the one in C# - Is this a problem?
-            Assert.AreEqual(s, "number: MyAccountNumber, sequence: MyAccountSequence, coins: System.Collections.Generic.List`1[sacco.Lib.StdCoin]");
+            Assert.AreEqual(s, "number: MyAccountNumber, sequence: MyAccountSequence, coins: System.Collections.Generic.List`1[commercio.sacco.lib.StdCoin]");
         }
     }
 }
