+++ sacco_test/utils/NodeInfoRetriever_test.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 using Newtonsoft.Json;
-using sacco.Lib;
+using commercio.sacco.lib;
 
 namespace sacco_test
 {
